+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - dynamic filter for quarters (e.g. %last quarter%)
+- Add Top N + others grouping #460
 
 ### Changed
 - JSON data source: remove column limit